requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.0
numpy>=1.26.0
Werkzeug>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
transformers==4.46.3
tokenizers==0.20.3
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0
addict>=2.4.0
matplotlib>=3.7.0
easydict>=1.9
einops>=0.7.0
easyocr>=1.7.0

